requirements.txt
flask==3.0.0
firebase-admin==6.2.0
numpy==1.24.3
pandas==2.0.3
yfinance==0.2.28
google-generativeai==0.3.1
flask-cors==4.0.0

